requirements.txt
agate==1.6.1
agate-dbf==0.2.2
agate-excel==0.2.3
agate-sql==0.5.5
asgiref==3.2.10
attrs==20.2.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
csvkit==1.0.5
dbfread==2.0.7
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
generics==2.0.0
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.0.1
isodate==0.6.0
isort==5.4.2
jdcal==1.4.1
leather==0.3.3
more-itertools==8.5.0
openpyxl==3.0.5
packaging==20.4
parsedatetime==2.6
#Pillow==8.1.0
pluggy==0.13.1
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.0
pytest-django==3.10.0
python-dotenv==0.14.0
python-slugify==4.0.1
pytimeparse==1.1.8
pytz==2020.1
requests==2.24.0
six==1.15.0
sorl-thumbnail==12.6.3
SQLAlchemy==1.3.19
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
transliterate==1.10.2
urllib3==1.25.10
wcwidth==0.2.5
xlrd==1.2.0
zipp==3.2.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
